Add files via upload
This dataset is an updated version of the student scores. I Concatenated the first and last names (full name). I also aggregated the scores in each course per student in another column (sum of scores). Afterwards, I used this updated dataset to create a pivot table and also plot a bar chart of the sum of scores against their first names. This bar chart clearly shows the student/students with the highest scores.